Feature/map2 (#217)
* 중간 push

* 중간점검

+++ R_DASH/src/main/java/com/pcwk/ehr/service/ShelterService.java
@@ -9,7 +9,7 @@ public interface ShelterService {
 	List<ShelterDTO> selectByBBox(double minLat, double maxLat, double minLon, double maxLon, String q, Integer limit)
 			throws Exception;
 
-	ShelterDTO selectOne(long shelterNo) throws Exception;
+	ShelterDTO selectOne(Integer shelterNo) throws Exception;
 
 	List<ShelterDTO> selectList(Integer limit) throws Exception;
